Move patches/ui sources to patches directory
