+++ UnleashedRecomp/CMakeLists.txt
@@ -129,15 +129,15 @@ set(UNLEASHED_RECOMP_HID_CXX_SOURCES
 )
 
 set(UNLEASHED_RECOMP_PATCHES_CXX_SOURCES
-    "patches/ui/CHudPause_patches.cpp"
-    "patches/ui/CTitleStateIntro_patches.cpp"
-    "patches/ui/CTitleStateMenu_patches.cpp"
-    "patches/ui/frontend_listener.cpp"
     "patches/aspect_ratio_patches.cpp"
     "patches/audio_patches.cpp"
     "patches/camera_patches.cpp"
     "patches/CGameModeStageTitle_patches.cpp"
+    "patches/CHudPause_patches.cpp"
+    "patches/CTitleStateIntro_patches.cpp"
+    "patches/CTitleStateMenu_patches.cpp"
     "patches/fps_patches.cpp"
+    "patches/frontend_listener.cpp"
     "patches/input_patches.cpp"
     "patches/inspire_patches.cpp"
     "patches/misc_patches.cpp"
