+++ UnleashedRecomp/patches/CHudPause_patches.cpp
