+++ UnleashedRecomp/patches/frontend_listener.cpp
