+++ UnleashedRecomp/patches/frontend_listener.h
